+++ components/cascader/interface.ts
@@ -1,3 +1,7 @@
 import type { OptionValue } from '../cascader-panel/interface';
 
 export type SelectValue = OptionValue | OptionValue[] | OptionValue[][];
+
+export type CascaderEmits = {
+    (e: 'removeTag', value: OptionValue): void;
+};
